Check to see if dir exists in build script

+++ testing/web3signer_tests/build.rs
@@ -82,7 +82,9 @@ pub async fn download_binary(dest_dir: PathBuf) {
     // Rename the web3signer directory so it doesn't include the version string. This ensures the
     // path to the binary is predictable.
     let web3signer_dir = dest_dir.join("web3signer");
-    fs::remove_dir_all(&web3signer_dir).unwrap();
+    if web3signer_dir.exists() {
+        fs::remove_dir_all(&web3signer_dir).unwrap();
+    }
     fs::rename(
         dest_dir.join(format!("web3signer-{}", version)),
         web3signer_dir,
